+++ src/main/java/com/mf/minutefictionbackend/controllers/CommentController.java
@@ -3,8 +3,11 @@ package com.mf.minutefictionbackend.controllers;
 import com.mf.minutefictionbackend.dtos.inputDtos.CommentInputDto;
 import com.mf.minutefictionbackend.dtos.outputDtos.CommentOutputDto;
 import com.mf.minutefictionbackend.services.CommentService;
+import com.mf.minutefictionbackend.services.SecurityService;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -15,13 +18,16 @@ import java.net.URI;
 public class CommentController {
 
     private final CommentService commentService;
-    public CommentController(CommentService commentService) {
+    private final SecurityService securityService;
+    public CommentController(CommentService commentService, SecurityService securityService) {
         this.commentService = commentService;
+        this.securityService = securityService;
     }
 
 
     @PostMapping("/stories/{storyId}/comments")
-    public ResponseEntity<CommentOutputDto> addCommentToStory(@Valid @PathVariable Long storyId, @RequestParam String username, @RequestBody CommentInputDto commentInputDto) {
+    public ResponseEntity<CommentOutputDto> addCommentToStory(@Valid @PathVariable Long storyId, @RequestBody CommentInputDto commentInputDto) {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
         CommentOutputDto savedComment = commentService.addComment(commentInputDto, storyId, username);
 
         URI uri = URI.create(ServletUriComponentsBuilder
@@ -31,14 +37,21 @@ public class CommentController {
         return ResponseEntity.created(uri).body(savedComment);
     }
 
+
     @PatchMapping("/comments/{commentId}")
     public ResponseEntity<CommentOutputDto> updateComment(@Valid @PathVariable("commentId") Long commentId, @RequestBody CommentInputDto updatedComment) {
+        if(!securityService.isCommentOwner(commentId)) {
+            throw new AccessDeniedException("You do not have permission to update this comment.");
+        }
         CommentOutputDto updatedCommentDto = commentService.updateComment(commentId, updatedComment);
         return ResponseEntity.ok().body(updatedCommentDto);
     }
 
     @DeleteMapping("/comments/{commentId}")
     public ResponseEntity<Void> deleteComment(@PathVariable("commentId") Long commentId) {
+        if(!securityService.isCommentOwnerOrEditor(commentId)) {
+            throw new AccessDeniedException("You do not have permission to delete this comment.");
+        }
         commentService.deleteCommentById(commentId);
         return ResponseEntity.noContent().build();
     }
