+++ src/main/java/com/mf/minutefictionbackend/controllers/UserController.java
@@ -5,15 +5,16 @@ import com.mf.minutefictionbackend.dtos.mappers.UserMapper;
 import com.mf.minutefictionbackend.dtos.outputDtos.UserOutputDto;
 import com.mf.minutefictionbackend.exceptions.BadRequestException;
 import com.mf.minutefictionbackend.models.User;
+import com.mf.minutefictionbackend.services.SecurityService;
 import com.mf.minutefictionbackend.services.UserService;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.net.URI;
-import java.util.Map;
 import java.util.Set;
 
 @RestController
@@ -21,10 +22,12 @@ import java.util.Set;
 public class UserController {
 
     private final UserService userService;
+    private final SecurityService securityService;
 
-    public UserController(UserService userService) {
+    public UserController(UserService userService, SecurityService securityService) {
         this.userService = userService;
 
+        this.securityService = securityService;
     }
 
     @PostMapping
@@ -41,29 +44,37 @@ public class UserController {
         return ResponseEntity.created(uri).body(userOutputDto);
     }
 
-    @PreAuthorize("hasAuthority('EDITOR')")
     @GetMapping
     public ResponseEntity<Set<UserOutputDto>> getAllUsers() {
+        if(!securityService.isEditor()) {
+            throw new AccessDeniedException("You do not have permission to perform this action.");
+        }
         return ResponseEntity.ok().body(userService.getAllUsers());
     }
 
-    @PreAuthorize("#username == authentication.principal.username || hasAuthority('EDITOR')")
     @GetMapping("/{username}")
     public ResponseEntity<UserOutputDto> getUserByUsername(@PathVariable("username") String username) {
+        if(!securityService.isOwnerOrEditor(username)) {
+            throw new AccessDeniedException("You do not have permission to see this users data.");
+        }
         UserOutputDto optionalUser = userService.getUserByUsername(username);
         return ResponseEntity.ok().body(optionalUser);
     }
 
-    @PreAuthorize("#username == authentication.principal.username || hasAuthority('EDITOR')")
     @DeleteMapping("/{username}")
     public ResponseEntity<Void> deleteUser(@PathVariable("username") String username) {
+        if(!securityService.isOwnerOrEditor(username)) {
+            throw new AccessDeniedException("You do not have permission to delete this user.");
+        }
         userService.deleteUser(username);
         return ResponseEntity.noContent().build();
     }
 
-    @PreAuthorize("#username == authentication.principal.username || hasAuthority('EDITOR')")
     @PutMapping("/{username}")
     public ResponseEntity<UserOutputDto> updateUser(@Valid @PathVariable("username") String username, @RequestBody UserInputDto userDto) {
+        if(!securityService.isOwner(username)) {
+            throw new AccessDeniedException("You do not have permission to update this users data.");
+        }
         UserOutputDto updatedUser = userService.updateUser(username, userDto);
         return ResponseEntity.ok().body(updatedUser);
     }
@@ -71,18 +82,21 @@ public class UserController {
 
     // MANAGE AUTHORITIES
 
-    @PreAuthorize("hasAuthority('EDITOR')")
     @GetMapping("/{username}/authorities")
     public ResponseEntity<Object> getUserAuthorities(@PathVariable("username") String username) {
+        if(!securityService.isEditor()) {
+            throw new AccessDeniedException("You do not have permission to request authorities.");
+        }
         return ResponseEntity.ok().body(userService.getAuthorities(username));
     }
 
-    @PreAuthorize("hasAuthority('EDITOR')")
     @PostMapping("/{username}/authority")
-    public ResponseEntity<Object> addUserAuthority(@PathVariable("username") String username, @RequestBody Map<String, Object> fields) {
+    public ResponseEntity<Object> addUserAuthority(@PathVariable("username") String username, @RequestBody String authority) {
+        if(!securityService.isEditor()) {
+            throw new AccessDeniedException("You do not have permission to add authorities.");
+        }
         try {
-            String authorityName = (String) fields.get("authority");
-            userService.addAuthority(username, authorityName);
+            userService.addAuthority(username, authority);
             return ResponseEntity.noContent().build();
         }
         catch (Exception ex) {
@@ -90,9 +104,11 @@ public class UserController {
         }
     }
 
-    @PreAuthorize("hasAuthority('EDITOR')")
     @DeleteMapping("/{username}/authorities/{authority}")
     public ResponseEntity<Object> deleteUserAuthority(@PathVariable("username") String username, @PathVariable("authority") String authority) {
+        if(!securityService.isEditor()) {
+            throw new AccessDeniedException("You do not have permission to delete authorities.");
+        }
         userService.removeAuthority(username, authority);
         return ResponseEntity.noContent().build();
     }
